Cumulative update. (#16)
* Cumulative update.

* Fix libpng reference in 3rdparty to point to a stable release.

* Updated Ubuntu build CI script to use Ubuntu 22.04 since 18.04 environment has been deprecated.

+++ src/utils/utl_image_2d.h
@@ -60,7 +60,7 @@ private:
   pixel_t*    m_data=nullptr;
   bool        m_is_owned=false;
   bool m_blocked = false;
-  int m_capacity_in_bytes = 0;
+  size_t m_capacity_in_bytes = 0;
 };
 
 }
